+++ CONTRIBUTING.md
@@ -23,4 +23,5 @@ In another terminal, you can for instance navigate to `examples/next-prisma-star
 yarn test --watch
 ```
 
-Testing is currently coalesced in [./packages/server/test](./packages/server/test) - we import the different libs from here, this makes it easier for us to do integration testing + getting test coverage on the whole codebase.
+All Integration testing is currently coalesced in [./packages/server/test](./packages/server/test) - we import the different libs from here, this makes it easier for us to do integration testing + getting test coverage on the whole codebase.
+
